feat: background video uploader
